#36 extracted submitOrder and edit callers with extra parameters
