+++ Project/Frontend/src/Order.jsx
@@ -10,7 +10,7 @@ import PaymentModal from './PaymentModal';
 import NewOrderModal from './NewOrderModal';
 import MenuCard from './MenuCard';
 import { Link } from 'react-router-dom';
-import { addItemToCart, replaceSuggestion, clearCart, removeItemFromCart } from './CartContext/cartUtils';
+import { addItemToCart, replaceSuggestion, clearCart, removeItemFromCart, submitOrder } from './CartContext/cartUtils';
 
 // Popup component to display order information
 const OrderInfoPopup = React.memo(({
@@ -163,7 +163,7 @@ OrderInfoPopup.propTypes = {
 
 function Order() {
     // Context and state variables
-    const { cart, fetchCart, menuItems, customer, submitOrder, createNewOrder, tableNum, setCart, setCustomer, setTableNum, suggestions, setSuggestions } = useContext(CartContext);
+    const { cart, fetchCart, menuItems, customer, createNewOrder, tableNum, setCart, setCustomer, setTableNum, suggestions, setSuggestions } = useContext(CartContext);
     const [message, setMessage] = useState(''); // snackbar message
     const [severity, setSeverity] = useState('success'); // snackbar severity
     const [showOrderInfo, setShowOrderInfo] = useState(() => { // Controls order info popup visibility
@@ -466,7 +466,7 @@ function Order() {
 
     // Handles payment success by submitting the order and updating the UI and local storage
     const handlePaymentSuccess = async () => {
-        const result = await submitOrder(); // submit order and get result 
+        const result = await submitOrder(customer, cart, ws); // submit order and get result 
         if (result.success) {
             // create a new order object 
             const newOrder = {
